requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
sqlalchemy==2.0.21
psycopg2-binary==2.9.7
python-multipart==0.0.6
python-dotenv==1.0.0
openai==1.3.7
pydantic==2.4.2
pydantic-settings==2.0.3
pydub==0.25.1
moviepy==1.0.3
ffmpeg-python==0.2.0
pytest==7.4.2
httpx==0.25.0 